Use fixed-width integers in file, packet and codec interfaces
Migrate `long` and `unsigned long` in the shared interfaces whose width
changed between Win32 and LP64:

- The FileClass chain (RawFileClass, BufferIOFileClass, RAMFileClass,
  both games' CCFileClass and TD's RawFileClass) and the Read_File family
  now take and return int32_t, casting at the int64_t POSIX boundary.
  The unused Get_Date_Time/Set_Date_Time virtuals and
  Transfer_Block_Size helpers are removed.
- Packet fields use int32_t/uint32_t. TYPE_LONG fields were sent as 8
  bytes on LP64; they are 4 bytes again, as in the original protocol.
- Blowfish, SHA, CRC straws and the multiprecision helpers use uint32_t;
  the trial-quotient intermediates in mp_quo_digit are uint64_t so they
  keep today's non-wrapping arithmetic. Buffer sizes are base::ssize.
- LCW_Uncompress, Buffer_Frame_To_Page and the WSA buffer size take
  int32_t; the Planet Westwood game ID and start time are uint32_t.

New tests cover packet field round trips and wire layout, and Blowfish
against the reference vectors.

+++ src/ra/2keyfram.cc
@@ -288,8 +288,7 @@ void* Build_Frame(const void* dataptr, const uint16_t framenumber,
     if (keyfr->flags & 1) {
       ptr = static_cast<char*>(Add_Long_To_Pointer(ptr, 768));
     }
-    length = static_cast<int32_t>(
-        LCW_Uncompress(ptr, buffptr, static_cast<unsigned long>(buffsize)));
+    length = LCW_Uncompress(ptr, buffptr, buffsize);
   } else {
     uint16_t currframe = 0;
     // key delta or delta
@@ -315,8 +314,7 @@ void* Build_Frame(const void* dataptr, const uint16_t framenumber,
       ptr = static_cast<char*>(Add_Long_To_Pointer(ptr, 768));
     }
 
-    length = static_cast<int32_t>(
-        LCW_Uncompress(ptr, buffptr, static_cast<unsigned long>(buffsize)));
+    length = LCW_Uncompress(ptr, buffptr, buffsize);
 
     if (length > buffsize) {
       return nullptr;
