+++ src/tech/crcpipe.cc
@@ -87,4 +87,4 @@ int CRCPipe::Put(const void* source, int slen) {
  *                                                                                             *
  * HISTORY: * 07/03/1996 JLB : Created. *
  *=============================================================================================*/
-long CRCPipe::Result() const { return crc_.Value(); }
+uint32_t CRCPipe::Result() const { return crc_.Value(); }
